You Win Text (Maybe)
I'm not sure what went wrong with the win text from the project version that was downloaded by others during the meeting. So just going to commit these changes I see on my computer which I have no idea what it's about. Maybe the You Win Text will be fixed with this?