Puzzle Manager handles puzzle scores. Timer issue fixed
Puzzle manager now holds all the puzzle score information and does update there, as opposed to doing it in the game manager.

Timer had a bug since the last few versions when there was an implementation to allow for turn taking and animation, where the player's defence scores were not counted, because it kept refreshing too early. Problem is now fixed, and scores are reflected accordingly while animation is taking turns.